KAFKA-3522: Add RocksDBTimestampedSegmentedBytesStore
