+++ streams/src/main/java/org/apache/kafka/streams/state/Stores.java
@@ -265,7 +265,13 @@ public class Stores {
                                                    + windowSize + "], retention=[" + retentionPeriod + "]");
         }
 
-        return new RocksDbWindowBytesStoreSupplier(name, retentionPeriod, segmentInterval, windowSize, retainDuplicates);
+        return new RocksDbWindowBytesStoreSupplier(
+            name,
+            retentionPeriod,
+            segmentInterval,
+            windowSize,
+            retainDuplicates,
+            false);
     }
 
     /**
